pyproject.toml
[project]
name = "semantic-kg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.9.3",
    "networkx>=3.4.2",
    "numpy>=2.1.3",
    "openai>=1.57.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "ruff>=0.8.2",
    "scipy>=1.14.1",
]
